create validator
