+++ configs/_base_/datasets/sun360_alpha_outdoor.py
@@ -9,6 +9,8 @@ dataset = dict(
     num_easy=30,
     num_medium=40,
     num_hard=30,
+    difficulty='easy',
+    bounded=False,
     fov=90.0,
     min_steps=10,
     max_steps=5000,
