+++ findview_baselines/rl/ppo/ppo.py
@@ -72,7 +72,7 @@ class PPO(nn.Module):
         action_loss_epoch = 0.0
         dist_entropy_epoch = 0.0
 
-        for _e in range(self.ppo_epoch):
+        for _ in range(self.ppo_epoch):
             data_generator = rollouts.recurrent_generator(
                 advantages, self.num_mini_batch
             )
